Remove dependency on Textract for Excel files
