+++ solution/text-extractor/extractors/__init__.py
@@ -12,3 +12,4 @@ from .binary import BinaryExtractor as BinaryExtractor
 from .email import EmailExtractor as EmailExtractor
 from .markup import MarkupExtractor as MarkupExtractor
 from .rtf import RichTextExtractor as RichTextExtractor
+from .excel import ExcelExtractor as ExcelExtractor
