+++ solution/text-extractor/extractors/office.py
@@ -7,9 +7,6 @@ from .extractor import Extractor
 class OfficeExtractor(Extractor):
     file_types = (
         'docx',
-        'xls',
-        'xlsx',
-        'xlsm',
         'pptx',
     )
