+++ solution/text-extractor/extractors/tests/fixtures/xlsm_expected.txt
@@ -1,43 +1,1 @@
-
-Report Version of 05/13/2021 
-Summary State MEQC Review Reports and Corrective Action Plans (CAPs)
-State Name: 
-Submission Date:
-MEQC Review Period: 
-State Contact Name 1:
-State Contact 1:  E-Mail Address
-State Contact 1:  Phone Number
-State Contact 1:  Correspondence Address
-State Contact Name 2:
-State Contact 2:  E-Mail Address
-State Contact 2:  Phone Number
-State Contact 2:  Correspondence Address
-Total Active Cases Reviewed:
-Total Medicaid Active Cases Reviewed (if available):
-Total CHIP Active Cases Reviewed (if available):
-Number of Negative Case Actions Reviewed:
-Total Medicaid Negative Case Actions Reviewed (if available):
-Total CHIP Negative Case Actions Reviewed (if available):
-Category of Reporting Data Reported
-Total number of cases reviewed (Break out by active cases and negative case actions):
-Percentage of correctly determined active cases:
-Percentage of correctly determined negative cases:
-Percentage of total correctly determined cases:
-Percentage of active cases with errors:
-Percentage of negative cases with errors:
-Percentage of total cases with errors:
-Percentage of active cases with technical deficiencies (TDs):
-Percentage of negative cases with technical deficiencies (TDs):
-Percentage of total cases with technical deficiencies (TDs):
-Top 10 frequent findings across all cases reviewed:                                                        (NOTE: Indicate how often each finding in the top 10 was cited--Example:                 1AF. Income incorrectly calculated [12]) 1.
-2.
-3.
-4.
-5.
-6.
-7.
-8.
-9.
-10.
-Summary CAP Report                                                   Summary CAP Report                                                  Summary CAP Report
-Top 10 Frequent Findings Across All Cases Reviewed                                      (Active and Negative)                                                                                         Root Cause of Error or Deficiency Type of Corrective Action Taken or To Be Taken Status of Corrective Action* Implementation Date Estimated Implementation Date Responsible Party How State Will Monitor Implementation of CAPs  How State Will Evaluate Effectiveness of CAP
+ Report Version of 05/13/2021  Summary State MEQC Review Reports and Corrective Action Plans (CAPs) State Name:  Submission Date: MEQC Review Period:  State Contact Name 1: State Contact 1:  E-Mail Address State Contact 1:  Phone Number State Contact 1:  Correspondence Address State Contact Name 2: State Contact 2:  E-Mail Address State Contact 2:  Phone Number State Contact 2:  Correspondence Address Total Active Cases Reviewed: Total Medicaid Active Cases Reviewed (if available): Total CHIP Active Cases Reviewed (if available): Number of Negative Case Actions Reviewed: Total Medicaid Negative Case Actions Reviewed (if available): Total CHIP Negative Case Actions Reviewed (if available): Category of Reporting Data Reported Total number of cases reviewed (Break out by active cases and negative case actions): Percentage of correctly determined active cases: Percentage of correctly determined negative cases: Percentage of total correctly determined cases: Percentage of active cases with errors: Percentage of negative cases with errors: Percentage of total cases with errors: Percentage of active cases with technical deficiencies (TDs): Percentage of negative cases with technical deficiencies (TDs): Percentage of total cases with technical deficiencies (TDs): Top 10 frequent findings across all cases reviewed:                                                        (NOTE: Indicate how often each finding in the top 10 was cited--Example:                 1AF. Income incorrectly calculated [12]) 1. 2. 3. 4. 5. 6. 7. 8. 9. 10. Summary CAP Report                                                   Summary CAP Report                                                  Summary CAP Report Top 10 Frequent Findings Across All Cases Reviewed                                      (Active and Negative)                                                                                         Root Cause of Error or Deficiency Type of Corrective Action Taken or To Be Taken Status of Corrective Action* Implementation Date Estimated Implementation Date Responsible Party How State Will Monitor Implementation of CAPs  How State Will Evaluate Effectiveness of CAP
