+++ solution/text-extractor/requirements.txt
@@ -10,3 +10,4 @@ extract-msg
 pillow>=10.0.1 # not directly required, pinned by Snyk to avoid a vulnerability
 beautifulsoup4
 striprtf
+xlrd
